refactor: format utils 파일 이름 변경
